Split tests from source files
I am making a distinction between source files and files
relevant for our package (tests, documentation, package
configuration, etc). It is the source files which are
imported into scripts and notebooks.

Most users aren't interested in installing documetation
or tests along with the source files, so these should be
placed in separate directories.