Prepare test suite with tox
I have added tox and pytest as optional dependencies of the package,
which can be installed through `pip install derivkit[test]`. Further
configuration of tox (which versions of python to test against,
where to run the tests from, where to store files) has been added
to pyproject.toml, and I have added some of the directories which
tox creates to .gitignore.