Move documentation from README to docs
A large part of the README is more appropriate for the landing
page of the main documentation. I have moved it to the index
page. I have also added an "assets" directory in the docs
directory to keep any PDFs, images or other things we want to
link to in the documentation. Probably most of the demos
currently in the notebooks can be placed here.

Probably this still needs to be further refined and rewritten.
At the very least, the docstrings are currently not proper RST.
This makes the documentation a bit awkward, but will have to be
changed incrementally.