Update docstrings to Google format
The module docstrings are written using various different
conventions. As a result, the documentation pages aren't always
properly rendered. I propose we rewrite the docstrings using
the Google docstring standard:

	https://google.github.io/styleguide/pyguide.html#docstrings

This makes use of the Napoleon module, which I had added previously:
(https://www.sphinx-doc.org/en/master/usage/extensions/napoleon.html).

In the case of the plotutils files I have only made the necessary
changes so that ruff doesn't complain. The reason for this being
that I think we discussed that these functions are mostly for our
own convenience and that general users aren't expected to import
them.