Add test if covariance_matrix is not a square numpy array
The __init__  of ForecastKit class states that covariance_matrix
must be square, but no exception is raised if this is not true.
Potentially this is an issue for get_forecast_tensors, as a
non-square matrix could still have a pseudo-inverse. I think it
is safer to catch this in the initialisation.