Add tests for ForecastKit
These are some simple tests for the ForecastKit. Unfortunately the
epxected output for get_forecast_tensors quickly becomes unyieldy
to put in the fixture directly. Currently I have combined 2 sets
of model functions and two sets of fiducial values-covariance matrices.
This results in 4 tests for this method.

I am sure there is a better way to do this, but this should be
something that is useful enough to get started. In the future it
may be interesting to see what other projects use for their testing
strategies (e.g. different libraries).