Fix the squareness check of the covariance matrix
The covariance matrix can be one-dimensional if there is a single
observable. However, the test in ForecastKit will fail. I have
split the test into two: one to ensure that the dimensionality is
at most two, and one to ensure that if the dimensionality is two
that the matrix is square.