Exclude attributes of ComponentOutcome from autodoc
Dataclasses have a special behaviour when attributes are specified
in the docstring. In this case, both Napoleon and Autodoc will declare
the attributes in the rST file, which results in a warning from Sphinx.
Because in our case warnings equal errors, this causes our Sphinx
deployment to fail.

I have found a solution in https://stackoverflow.com/a/64601933.
The suggestion seems to be to either turn off Autodoc's parsing,
or to explicitly exclude the attributes in the relevant rST file.
Since we currently only have 1 dataclass, I opted for the latter.